Refactor codes
Change the function 'open' and 'read' used to process the file
information with 'fopen' and 'fgets' respectively due to the chanllenges
faced with identification of a line where only a new line occurs.
The newline used as a delimiter after the whole file is read is making
it difficult to know where only a newline is passed.